feat: implement production calculation with BigDecimal and tests
- Add production optimization logic in ProductionOptimizerService
- Replace usage of double with BigDecimal for higher precision
- Create unit tests to validate results and business rules
- Add input validations in services and DTOs
- General adjustments in controllers and services for compatibility

+++ src/main/java/com/productionoptimizer/dto/ProductDTO.java
@@ -3,6 +3,7 @@ package com.productionoptimizer.dto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -10,5 +11,5 @@ import lombok.AllArgsConstructor;
 public class ProductDTO {
     private Long id;
     private String name;
-    private double salePrice;
+    private BigDecimal salePrice;
 }
